% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor_cs}
\alias{cor_cs}
\title{Create a compound-symmetric correlation matrix}
\usage{
cor_cs(r, m = 2)
}
\arguments{
\item{r}{rho, correlation coefficient (numeric/scalar)}

\item{m}{number of rows/columns (numeric/scalar)}
}
\description{
Create a compound-symmetric correlation matrix
}
