% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-data_as_code.R
\name{data_as_code}
\alias{data_as_code}
\title{Convert data.frame to R code that creates identical dataset}
\usage{
data_as_code(data)
}
\arguments{
\item{data}{data.frame to convert (data.frame)}
}
\value{
object containing code to generate data.frame (character/scalar)
}
\description{
Convert data.frame to R code that creates identical dataset
}
\note{
TODO: handle missing data (currently using datapasta pkg instead for this)
}
