% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-gst_power.R
\name{esWidgetServer}
\alias{esWidgetServer}
\title{Sidebar Effect Size Server}
\usage{
esWidgetServer(id)
}
\arguments{
\item{id}{namespaced ID (character/scalar)}
}
\description{
Sidebar Effect Size Server
}
