% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-gst_power.R
\name{esWidgetUI}
\alias{esWidgetUI}
\title{Sidebar Effect Size UI}
\usage{
esWidgetUI(id)
}
\arguments{
\item{id}{namespaced ID (character/scalar)}
}
\description{
This module is for computing GST Power for >=2 joint
endpoints based on O-Brien OLS
}
