% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-forest_data.R
\name{forest_data}
\alias{forest_data}
\title{Forestplot Example Data}
\usage{
forest_data()
}
\value{
A data.frame that can be passed to \code{\link{forest_plot}}
}
\description{
Forestplot Example Data
}
