% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-forest_import_csv.R
\name{forest_import_csv}
\alias{forest_import_csv}
\title{Import forest plot data from csv file}
\usage{
forest_import_csv(filename)
}
\arguments{
\item{filename}{full path to csv file (character/scalar)}
}
\description{
Import forest plot data from csv file
}
