% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-forest_reprex.R
\name{forest_reprex}
\alias{forest_reprex}
\title{Create reproducible code for current forestplot}
\usage{
forest_reprex(forestplot, unfold = FALSE)
}
\arguments{
\item{forestplot}{output from forest_plot()}

\item{unfold}{unfold the \code{\link{forest_plot}} function? (logical/scalar; default = FALSE)}
}
\description{
Create reproducible code for current forestplot
}
\note{
Unfolding the code (\code{unfold = T}) will generate a code snippet that uses the
\code{\link{forestploter}} package functions directly; the resulting code
does not require the \code{gst} package.

WARNING: This function will not create a perfect reprex if you use the \code{...} argument in \code{\link{forest_plot}}
}
