% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-forest_suggest_params.R
\name{forest_suggest_params}
\alias{forest_suggest_params}
\title{Suggest parameters for a forest_plot}
\usage{
forest_suggest_params(dt)
}
\arguments{
\item{dt}{forest data}
}
\description{
Suggest parameters for a forest_plot
}
