% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrien.R
\name{get_n}
\alias{get_n}
\title{Calculate sample size}
\usage{
get_n(
  power,
  es0,
  es1,
  R,
  alpha = 0.05,
  k,
  method = c("ols", "gls"),
  df_method = c("vd", "vob", "vlt"),
  n0_range = c(2, 1000),
  tolerance = 1e-06
)
}
\arguments{
\item{power}{target power}

\item{es0}{effect size in endpoint 1}

\item{es1}{effect size in endpoint 2}

\item{R}{correlation matrix between endpoints}

\item{alpha}{significance level}

\item{k}{sample size ratio of treatment vs placebo}

\item{method}{obrien "ols",  "gls"}

\item{df_method}{degrees-of-freedom method (one of: {vd, vob, vlt})}

\item{n0_range}{range of size in placebo to be searched}

\item{tolerance}{the desired accuracy (convergence tolerance)}
}
\value{
power
}
\description{
Calculate sample size
}
