% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrien.R
\name{get_oes}
\alias{get_oes}
\title{Calculate operational effect size}
\usage{
get_oes(lambda, R, method = c("ols", "gls"))
}
\arguments{
\item{lambda}{vector of effect size for each endpoint}

\item{R}{correlation matrix of lambda}

\item{method}{obrien "ols",  "gls"}
}
\value{
A scalar
}
\description{
Calculate operational effect size
}
