% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrien.R
\name{get_power}
\alias{get_power}
\title{Calculate Power}
\usage{
get_power(oes, n0, n1 = n0, df = n0 + n1 - 2, alpha = 0.05)
}
\arguments{
\item{oes}{operational effect size}

\item{n0}{sample size in placebo group}

\item{n1}{sample size in treatment group}

\item{df}{degree of freedom}

\item{alpha}{significance level}
}
\value{
power
}
\description{
Calculate Power
}
