% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GST.parameter.R
\name{GST.parameter}
\alias{GST.parameter}
\title{GST.parameter function}
\usage{
GST.parameter(data, n1)
}
\arguments{
\item{data}{N-by-k matrix whose first n1 rows are observations from group 1}

\item{n1}{number of observations in group 1}
}
\value{
unbiased estimator of theta

estimated standard deviation of GTE

estimated S1.n1n2

estimated S2.n1n2

estimated S3.n1n2

unbiased estimator of theta sigma

estimated h1, adjusted rank-sum-type test

estimated h2, adjusted rank-sum-type test
}
\description{
This function calculate GST p.value and estimate GTE
}
