% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GST.R
\name{GST}
\alias{GST}
\title{GST function}
\usage{
GST(data, m, test.side)
}
\arguments{
\item{data}{a matrix of raw data. Rows are subjects and columns are endpoints.
All columns must be coded so that larger value is a better outcome.
The first m rows in data are data from treatment 1, and the rest rows are data from treatment 2}

\item{m}{number of subjects in treatment 1}

\item{test.side}{1(one-sided test) or 2 (two-sided test)}
}
\value{
\code{p.value}: p value of the global statistical test

\code{h}: adjusting factor to O'Brien's test

\code{theta.hat}: a vector of treatment effect on individual outcomes

\code{GTE}: global treatment effect of all outcomes

\code{sigmaGTE}: standard deviation of GTE estimate
}
\description{
This function calculates one or two-sided non-parametric global statistical test
}
