% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GST.simu.R
\name{GST.simu}
\alias{GST.simu}
\title{GST.simu function}
\usage{
GST.simu(data, m)
}
\arguments{
\item{data}{N-by-k matrix whose first n1 rows are observations from group 1}

\item{m}{number of observations in group 1}
}
\value{
p value of the global statistical test

global treatment effect of all outcomes

p value under univariate assumption
}
\description{
This function performs simulation using GST
}
