% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gst_power_app.R
\name{gst_power_app}
\alias{gst_power_app}
\title{Launch the GST Power App}
\usage{
gst_power_app(runApp = TRUE)
}
\arguments{
\item{runApp}{should the app be run on completion, default is TRUE}
}
\value{
A Shiny app
}
\description{
Launch the GST Power App
}
\examples{
if (interactive()) {
  gst_power_app()
}
}
