% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod-gst_tipping_point.R
\name{gstTippingPoint_ui}
\alias{gstTippingPoint_ui}
\title{Effect size tipping point calculation}
\usage{
gstTippingPoint_ui(id)
}
\arguments{
\item{id}{namespaced ID (character/scalar)}
}
\description{
Given Endpoint 1 effect size ES1 and correlation with endpoint 2 Rho, calculate tipping point for Effect Size 2, such that the joint endpoint effect size dips
below the first (larger) Endpoint 1 ES1. This is based on Dallow (2016)
}
