% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forestplot_app.R
\name{make_forestplot_app}
\alias{make_forestplot_app}
\title{A Shiny app to make a forest plot}
\usage{
make_forestplot_app(runApp = TRUE, plot_width = 1000, plot_height = 500)
}
\arguments{
\item{runApp}{should the app be run on completion, default is TRUE}

\item{plot_width}{Width of the plot in pixels, default is 1200}

\item{plot_height}{Height of the plot in pixels, default is 800}
}
\value{
A Shiny app
}
\description{
A Shiny app to make a forest plot
}
\examples{
if (interactive()) {
  make_forestplot_app()
}
}
