% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.GSTnoData.R
\name{N.GSTnoData}
\alias{N.GSTnoData}
\title{N.GSTnoData function}
\usage{
N.GSTnoData(k, r0, theta, sigma2, rho, alpha, power, test.side)
}
\arguments{
\item{k}{number of endpoints}

\item{r0}{n2/n1 randomization ratio}

\item{theta}{theta is a positive GTE where power is controlled}

\item{sigma2}{upper bound of variances F1(X2) and F2(X1)}

\item{rho}{upper bound of absolute correlation between F1(X2) and F2(X1)}

\item{alpha}{type I error}

\item{power}{desired statistical power}

\item{test.side}{1(one-sided test) or 2 (two-sided test)}
}
\value{
\code{n1}: sample size needed from treatment 1

\code{n2}: sample size needed from treatment 2

\code{N}: total sample size
}
\description{
This function calculates sample size needed in each group when there is no pilot data
}
\details{
This function calculates sample size based on information from literature.
In most cases, such information is presented in terms of mean and standard
deviation using normal approximation.
}
