% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.GSTwPilotData.R
\name{N.GSTwPilotData}
\alias{N.GSTwPilotData}
\title{N.GSTwPilotData function}
\usage{
N.GSTwPilotData(pilot.data, m1, r0, test.side, alpha, power, GTE)
}
\arguments{
\item{pilot.data}{a (m1+m2)xk matrix of pilot data with its}

\item{m1}{number of patients in treatment 1 in the pilot data}

\item{r0}{n2/n1, randomization ratio in new study}

\item{test.side}{1(one-sided test) or 2 (two-sided test)}

\item{alpha}{type I error to be controlled}

\item{power}{statistical power to be controlled}

\item{GTE}{global treatment effect where power is controlled at.}
}
\value{
sample size needed from treatment 1

sample size needed from treatment 2

unbiased estimate of GTE from the pilot data
}
\description{
This function calculate sample size when some pilot data are available
}
