% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrien.R
\name{obrien}
\alias{obrien}
\title{O’Brien test}
\usage{
obrien(
  data,
  endpoint_vars,
  treatment_vars,
  method = c("ols", "gls", "ranksum"),
  var.equal = TRUE,
  ...
)
}
\arguments{
\item{data}{dataset with treatment and multiple outcomes in wide format}

\item{endpoint_vars}{the character vectors of outcomes in the dataset dat}

\item{treatment_vars}{treatment varialbe in the dataset dat}

\item{method}{obrien "ols",  "gls" and  "ranksum" test}

\item{var.equal}{default TRUE. the obrien paper assumes equal variance and using pooled variance to standardize data}

\item{...}{other options passed to \code{\link[stats]{t.test}} and \code{\link[stats]{wilcox.test}}}
}
\value{
A list with t-statistic, df, pvalue, alternative, and correlation matrix R
}
\description{
O’Brien test
}
\examples{

# prepare example data (from multcomp::mtept)
dat <- multcomp::mtept
dat$E4 <- -dat$E4  # align signs for all endpoint

# OLS method
dat |>
   obrien(
     endpoint_vars = c("E1", "E2", "E3", "E4"),
     treatment_vars = "treatment",
     method = "ols"
     )

# GLS one-sided
dat |>
  obrien(
    endpoint_vars = c("E1", "E2", "E3"),
    treatment_vars = "treatment",
    method = "gls",
    alternative = "less"
    )

# Ranksum method
dat |>
  obrien(
    endpoint_vars = c("E1", "E2", "E3", "E4"),
    treatment_vars = "treatment",
    method = "ranksum"
    )

}
