% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{p_combine}
\alias{p_combine}
\title{P-value combination method}
\usage{
p_combine(p, R, method = c("fisher", "stouffer"))
}
\arguments{
\item{p}{p-values (numeric/vector)}

\item{R}{correlation matrix}

\item{method}{choose between 'fisher' or 'stouffer' methods}
}
\description{
P-value combination method
}
