% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmnv}
\alias{rmnv}
\title{Simulate from a Multivariate Normal Distribution}
\usage{
rmnv(n, mean_v, cor_mtx, treatment = "trt", ...)
}
\arguments{
\item{n}{number of samples}

\item{mean_v}{vector of means}

\item{cor_mtx}{correlation matrix for input variables}

\item{treatment}{treatment column value (character/scalar)}

\item{...}{additional parameters passed to \code{\link[MASS]{mvrnorm}}}
}
\description{
Simulate from a Multivariate Normal Distribution
}
