% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obrien.R
\name{sim_power}
\alias{sim_power}
\title{Calculate Power by simulation}
\usage{
sim_power(
  n0,
  n1 = n0,
  lambda,
  R,
  alpha = 0.05,
  N = 1000,
  ncore = 4,
  method = c("ols", "gls"),
  ...
)
}
\arguments{
\item{n0}{sample size in placebo group}

\item{n1}{sample size in treatment group}

\item{lambda}{vector of effect size for each endpoint}

\item{R}{correlation matrix of lambda}

\item{alpha}{significance level}

\item{N}{number of replication}

\item{ncore}{n of cores planned}

\item{method}{obrien "ols", "gls"}

\item{...}{additional arguments passed to \code{\link{obrien}}}
}
\value{
power
}
\description{
Calculate Power by simulation
}
