% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{theta}
\alias{theta}
\title{theta function}
\usage{
theta(x1, x2)
}
\arguments{
\item{x1}{observations from group 1}

\item{x2}{observations from group 2}
}
\value{
\code{theta1}: unbiased estimator of theta, which is the treatment effect on individual outcomes
}
\description{
This function calculates global treatment effect
}
