% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-tooltips.R
\name{tooltip_label}
\alias{tooltip_label}
\title{Custom label with optional tooltip icons}
\usage{
tooltip_label(label, info = NULL, alert = NULL)
}
\arguments{
\item{label}{text label (character/scalar)}

\item{info}{info hover tooltip text (character/scalar)}

\item{alert}{alert hover tooltip text (character/scalar)}
}
\description{
Custom label with optional tooltip icons
}
