% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GST.parameter.R
\name{GST.parameter}
\alias{GST.parameter}
\title{GST.parameter function}
\usage{
GST.parameter(data, n1)
}
\arguments{
\item{data}{N-by-k matrix whose first n1 rows are observations from group 1}

\item{n1}{number of observations in group 1}
}
\value{
\code{theta.hat}: an unbiased estimator the treatment effect on individual outcomes

\code{sigmaGTE}: an unbiased estimate of standard deviation of GTE

\code{S1.n1n2}: estimated S1.n1n2

\code{S1.n1n2}: estimated S2.n1n2

\code{S1.n1n2}: estimated S3.n1n2

\code{Sigma}: unbiased estimate of standard deviation of treatment effect on individual outcomes

\code{h1}: estimated h1, adjusted rank-sum-type test

\code{h2}: estimated h2, adjusted rank-sum-type test
}
\description{
This function provides unbiased estimators of GST parameters for public sharing,
which includes treatment effect on individual outcomes, standard deviation of GTE and etc.
}
