% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GST.R
\name{GST}
\alias{GST}
\title{GST function}
\usage{
GST(data, m, test.side)
}
\arguments{
\item{data}{a matrix of raw data. Rows are subjects}

\item{m}{number of subjects in treatment 1}

\item{test.side}{1(one-sided test) or 2 (two-sided test)}
}
\value{
p value of the global statistical test

adjusting factor to O'Brien's test

a vector of treatment effect on individual outcomes

global treatment effect of all outcomes

standard deviation of GTE estimate
}
\description{
This function calculates one or two-sided non-parametric global statistical test
}
