% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-obt.R
\name{obt}
\alias{obt}
\title{Perform Obrien truncated closed testing}
\usage{
obt(
  data,
  endpoint_vars,
  jendpoint = NULL,
  treatment_vars,
  alpha = 0.05,
  nu = 0.8,
  method = c("ols", "gls", "ranksum"),
  var.equal = TRUE,
  ...
)
}
\arguments{
\item{data}{input data; columns should contain endpoint data and must be named E1, E2}

\item{endpoint_vars}{vector of endpoint column names (character/vector)}

\item{jendpoint}{joint endpoint, a vector of individual endpoints (character/vector)}

\item{treatment_vars}{treatment column name (character/scalar)}

\item{alpha}{significance level (numeric/scalar)}

\item{nu}{truncation parameter, between 0 and 1 (numeric/scalar; default = 0.8)}

\item{method}{estimation method (character/scalar; default = 'OLS')}

\item{var.equal}{assume equal variances? (logical/scalar; default = TRUE)}

\item{...}{additional arguments passed to \code{\link{obrien}}}
}
\value{
a list containing the following elements:
\itemize{
\item \code{F}: target family of hypotheses (character/vector)
\item \code{df}: data.frame with relevant hypotheses and parameters
\item \code{df_gst}: data.frame with results from the different procedures
\item \code{alpha_A}: unused alpha available for propagation to the next family
}
}
\description{
Perform Obrien truncated closed testing
}
\note{
This is the truncated closed test as proposed by:
Luo, X., Li, L., Savenkov, O., Liu, W., Ni, X., Tang, W. and Guo, W., 2025.
Multiple Comparisons Procedures for Analyses of Joint Primary Endpoints and
Secondary Endpoints. Pharmaceutical Statistics, 24(3), p.e70010
}
\section{Input Data Requirements}{

The endpoint variables in \code{data} must be named \code{E1}, \code{E2}, \code{E3}, etc. for
the function to work correctly.  If you endpoint columns do not have this
naming convention, you must rename them.
}

\examples{
dat <- multcomp::mtept
dat$E4 <- -dat$E4  # align signs for all endpoints
obt(
    data = dat,
    endpoint_vars = c("E1","E2","E3","E4"),
    jendpoint = c("E3", "E4"),
    alpha = 0.05,
    nu = 0.6,
    treatment_vars = "treatment",
    method = "ols"
)

}
